feat: permit user deletion to himself only

+++ src/api/v1/endpoints/users.py
@@ -1,20 +1,18 @@
 from sqlalchemy.orm.session import Session
-from fastapi import APIRouter, Depends, HTTPException, status, Response
+from fastapi import APIRouter, Depends, status, Response
 from pydantic import parse_obj_as
 
 from dependencies import get_db_session, get_current_user
 from crud import user
 from schemas.user import UserInDbSchema, PublicUserSchema
+from utils import exceptions
 
 
 router = APIRouter()
 
 
 @router.get("/profile", response_model=PublicUserSchema, status_code=status.HTTP_200_OK)
-def get_profile(
-    current_user: UserInDbSchema = Depends(get_current_user),
-    session: Session = Depends(get_db_session),
-):
+def get_profile(current_user: UserInDbSchema = Depends(get_current_user)):
     return current_user
 
 
@@ -41,7 +39,7 @@ def get_user_by_id(
 ):
     requested_user = user.get_user_by_id(session, user_id)
     if not requested_user:
-        raise HTTPException(status_code=404, detail="user-not-found")
+        raise exceptions.NotFound(detail="user-not-found")
 
     return parse_obj_as(PublicUserSchema, requested_user)
 
@@ -54,7 +52,9 @@ def delete_user_by_id(
 ):
     requested_user = user.get_user_by_id(session, user_id)
     if not requested_user:
-        raise HTTPException(status_code=404, detail="user-not-found")
+        raise exceptions.NotFound(detail="user-not-found")
+    if requested_user.id != current_user.id:
+        raise exceptions.PermissionDenied(detail="can-not-delete-other-user")
 
     user.delete_user(session, requested_user)
     return Response(status_code=status.HTTP_204_NO_CONTENT)
